Rename RandomStringBuilder and make it immutable and thread-safe.
The RandomStringBuilder class was renamed to RandomStringGenerator. The
class is now constructed using an inner Builder class, resulting in an
immutable and thread-safe generator. The default random generator is now
ThreadLocalRandom rather than Random.

The inner CodePointPredicate class was extracted into a separate class
and renamed to CharacterPredicate. The commonly used predicates are now
in a separate enum, CharacterPredicates.