[TEXT-212] A more efficient implementation for calculating LCS
1. Additions:
(a) An implementation of Hirschberg's Longest Commons Substring (LCS) algorithm. This implementation is more space efficient than the previous version (linear vs. quadratic), but the time complexity stays the same.
(b) Javadoc and comments are updated accordingly and a couple of typos are fixed.
(c) JMH performance analysis for LongestCommonSubsequence
(d) A CSV file containing test inputs for JMH analysis is added to resources. Please note that jaring this text file (which contains only English characters) should make it quite small. So, we should not worry about the JAR size.
(e) Modified POM file: makred Ali Ghanbari as a contributor and made several modifications to avoid conflict.

2. Modifications:
(a) The method longestCommonSubstringLengthArray is marked as deprecated.